+++ cohttp-eio/tests/test_client.t
@@ -7,7 +7,9 @@ Test Client.get
   meth: GET
   resource: /get
   version: HTTP/1.1
-  headers: Header { Accept = "application/json"; Host = "localhost:8082" }
+  headers: Header {
+   Accept = "application/json"; Host = "localhost:8082";
+   User-Agent = "cohttp-eio" }
 
   $ kill ${running_pid}
 
@@ -21,8 +23,8 @@ Test Client.post
   resource: /post
   version: HTTP/1.1
   headers: Header {
-   Accept = "application/json"; Content-Length = "12"; Host = "localhost:8082"
-   }
+   Accept = "application/json"; Content-Length = "12"; Host = "localhost:8082";
+   User-Agent = "cohttp-eio" }
   
   hello world!
 
@@ -39,7 +41,8 @@ Test posting "chunked" data
   version: HTTP/1.1
   headers: Header {
    Content-Length = "23"; Header1 = "Header1 value text";
-   Content-Type = "text/plain"; Host = "localhost:8082" }
+   Content-Type = "text/plain"; Host = "localhost:8082";
+   User-Agent = "cohttp-eio" }
   
   size: 7
    data: Mozilla